feat: add moviesByQuery
