+++ src/utils/moviesByQuery.js
@@ -1,0 +1,9 @@
+export const findMovies = (movies, query) => {
+  const moviesByQuery = movies.filter((movie) => {
+    const movieRu = String(movie.nameRU).toLowerCase().trim();
+    const movieEn = String(movie.nameEN).toLowerCase().trim();
+    const userQuery = query.toLowerCase().trim();
+    return movieRu.indexOf(userQuery) !== -1 || movieEn.indexOf(userQuery) !== -1;
+  });
+  return moviesByQuery;
+};
